增加aop、swagger 校验等功能

+++ src/main/java/com/example/fundsystembackend/service/impl/MyFundServiceImpl.java
@@ -1,14 +1,21 @@
 package com.example.fundsystembackend.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.example.fundsystembackend.entity.ChinaMutualFundDescription;
 import com.example.fundsystembackend.entity.FundUser;
 import com.example.fundsystembackend.entity.MyFund;
+import com.example.fundsystembackend.mapper.ChinamutualfunddescriptionMapper;
+import com.example.fundsystembackend.mapper.ChinamutualfundnavMapper;
+import com.example.fundsystembackend.mapper.FunduserMapper;
 import com.example.fundsystembackend.mapper.MyfundMapper;
 import com.example.fundsystembackend.service.MyFundService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.example.fundsystembackend.utils.result.ApiResult;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * <p>
@@ -24,6 +31,15 @@ public class MyFundServiceImpl extends ServiceImpl<MyfundMapper, MyFund> impleme
     @Autowired
     private MyfundMapper myfundMapper;
 
+    @Autowired
+    private FunduserMapper funduserMapper;
+
+    @Autowired
+    private ChinamutualfunddescriptionMapper chinamutualfunddescriptionMapper;
+
+    @Autowired
+    private ChinamutualfundnavMapper chinamutualfundnavMapper;
+
     @Override
     public ApiResult addFund(String fundId, String userId) {
         if(fundId == null || fundId.equals("")) {
@@ -31,6 +47,19 @@ public class MyFundServiceImpl extends ServiceImpl<MyfundMapper, MyFund> impleme
         }
         if(userId == null || userId.equals("")){
             return ApiResult.error(1201, "userId不能为空");
+        }
+        FundUser fundUser = funduserMapper.selectById(userId);
+        if(fundUser==null) {
+            return ApiResult.error(1202, "该用户不存在");
+        }
+        ChinaMutualFundDescription chinaMutualFundDescription = chinamutualfunddescriptionMapper.selectById(fundId);
+        if(chinaMutualFundDescription == null){
+            return ApiResult.error(1202, "该基金不存在");
+        }
+        List<MyFund> mySelectedFundList = myfundMapper.selectList(
+                new QueryWrapper<MyFund>().select("fund_id", fundId).eq("user_id", userId));
+        if(mySelectedFundList != null && mySelectedFundList.size() > 0) {
+            return ApiResult.error(1203, "重复选择");
         }
         MyFund myFund = new MyFund();
         myFund.setFundId(fundId);
@@ -47,8 +76,58 @@ public class MyFundServiceImpl extends ServiceImpl<MyfundMapper, MyFund> impleme
         if(userId == null || userId.equals("")){
             return ApiResult.error(1201, "userId不能为空");
         }
-        myfundMapper.delete(new QueryWrapper<MyFund>().eq("", ""));
-        return ApiResult.success();
+        FundUser fundUser = funduserMapper.selectById(userId);
+        if(fundUser==null) {
+            return ApiResult.error(1202, "该用户不存在");
+        }
+        ChinaMutualFundDescription chinaMutualFundDescription = chinamutualfunddescriptionMapper.selectById(fundId);
+        if(chinaMutualFundDescription == null){
+            return ApiResult.error(1202, "该基金不存在");
+        }
+        List<MyFund> mySelectedFundList = myfundMapper.selectList(
+                new QueryWrapper<MyFund>().select("fund_id", fundId).eq("user_id", userId));
+        if(mySelectedFundList != null && mySelectedFundList.size() == 0) {
+            myfundMapper.delete(new QueryWrapper<MyFund>().eq("", ""));
+            return ApiResult.success();
+        }else {
+            return ApiResult.error(1203, "删除自选失败");
+        }
+
     }
 
+    @Override
+    public ApiResult fundList(String userId) {
+        if (userId == null || userId.equals("")) {
+            return ApiResult.error(1201, "userId不能为空");
+        }
+        FundUser fundUser = funduserMapper.selectById(userId);
+        if (fundUser == null) {
+            return ApiResult.error(1202, "该用户不存在");
+        }
+        List<MyFund> mySelectedFundList = myfundMapper.selectList(
+                new QueryWrapper<MyFund>().eq("user_id", userId));
+        if (mySelectedFundList == null || mySelectedFundList.size() == 0) {
+            return ApiResult.success(mySelectedFundList);
+        }
+        List<ChinaMutualFundDescription> chinaMutualFundDescriptions = new ArrayList<>();
+        for (MyFund myFund : mySelectedFundList) {
+            ChinaMutualFundDescription myFundDesc = chinamutualfunddescriptionMapper.selectById(
+                    myFund.getId()
+            );
+            chinaMutualFundDescriptions.add(myFundDesc);
+        }
+        for (ChinaMutualFundDescription chinaMutualFundDescription : chinaMutualFundDescriptions) {
+            String fInfoCode = chinaMutualFundDescription.getFInfoWindCode();
+
+            // List<> chinamutualfundnavList =
+            //         chinamutualfundnavMapper.selectList(
+            //                 new QueryWrapper<Chinamutualfundnav>().eq("f_info_windcode", fInfoCode)
+            //         );
+            // System.out.println(chinamutualfundnavList);
+            // if (chinamutualfundnavList != null && chinamutualfundnavList.size() > 0) {
+            //     chinaMutualFundDescription.setNewNavUnit() = chinamutualfundnavList.get(0).getFNavUnit();
+            // }
+        }
+        return ApiResult.success(chinaMutualFundDescriptions);
+    }
 }
